+++ lib/src/command_runner.dart
@@ -1,5 +1,6 @@
 import 'package:args/command_runner.dart' as args;
 import 'package:dart_dependency_checker_cli/src/deps_unused/deps_unused_command.dart';
+import 'package:dart_dependency_checker_cli/src/deps_used/deps_used_command.dart';
 import 'package:dart_dependency_checker_cli/src/transitive_use/transitive_use_command.dart';
 
 const versionFlag = 'version';
@@ -15,6 +16,7 @@ final class CommandRunner extends args.CommandRunner<int> {
       negatable: false,
       help: 'Print this package version.',
     );
+    addCommand(DepsUsedCommand());
     addCommand(DepsUnusedCommand());
     addCommand(TransitiveUseCommand());
   }